feat: #4 설문 관련 API

+++ src/main/java/com/yourmode/yourmodebackend/domain/survey/repository/SurveyAnswerRepository.java
@@ -3,5 +3,5 @@ package com.yourmode.yourmodebackend.domain.survey.repository;
 import com.yourmode.yourmodebackend.domain.survey.entity.SurveyAnswer;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SurveyAnswerRepository extends JpaRepository<SurveyAnswer, Long> {
+public interface SurveyAnswerRepository extends JpaRepository<SurveyAnswer, Integer> {
 }
